Almost Done

+++ src/app/registrar/registrar.component.ts
@@ -9,34 +9,65 @@ import Swal from 'sweetalert2';
 })
 export class RegistrarComponent {
 
-  constructor(
-    public router:Router
-    ){
-  
-    }
+  constructor(public router:Router){}
 
-  users: any[] = []
-  userSignUpObject: any = {
+  userSignUpObject: any = [{
     username: '',
     email: '',
     password: '',
-  }
+  }]
 
-  
+  users: any = [this.userSignUpObject];
+
+
   onSignUp(){
-    this.users.push(this.userSignUpObject)
-    localStorage.setItem('users', JSON.stringify(this.users))
 
-    Swal.fire({
-      title: "User created!",
-      icon: "success"
-    });
+    const userData = localStorage.getItem('users');
+    const u = JSON.stringify(userData)
+    const uP = JSON.parse(u)
 
-    this.userSignUpObject = {
-      username: '',
-      email: '',
-      password: '',
+
+    if(this.userSignUpObject.username == '' && this.userSignUpObject.email == '' && this.userSignUpObject.password == ''){
+      Swal.fire({
+        title: "Fields empty",
+        icon: "error"
+      });
+    }else if(!this.userSignUpObject.username == null || this.userSignUpObject.username == undefined || this.userSignUpObject.username == ''){
+      Swal.fire({
+        title: "Username field empty",
+        icon: "info"
+      });
+    }else if(!this.userSignUpObject.email == null || this.userSignUpObject.email == undefined || this.userSignUpObject.email == ''){
+      Swal.fire({
+        title: "Email field empty",
+        icon: "info"
+      });
+    }else if(!this.userSignUpObject.password == null || this.userSignUpObject.password == undefined || this.userSignUpObject.password == ''){
+      Swal.fire({
+        title: "Password field empty",
+        icon: "info"
+      });
+    }else if (this.userSignUpObject == uP) {
+        Swal.fire({
+          title: "The user already exists!",
+          icon: "error"
+        });
+    }else{
+      this.users.push(this.userSignUpObject)
+      localStorage.setItem('users', JSON.stringify(this.users))
+      Swal.fire({
+        title: "User created!",
+        icon: "success"
+      });
+
+      // this.userSignUpObject = {
+      //   username: '',
+      //   email: '',
+      //   password: '',
+      // }
     }
+    console.log(uP)
+    console.log(this.userSignUpObject)
   }
 
   goToSignIn(){
